backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
yfinance>=0.2.36
pandas>=2.0.0
finnhub-python>=2.4.19
python-dotenv>=1.0.0
